fix(gates): split the I/O layer out, and grant the scope the label bootstrap needs (#3726)
Two review findings from Codex, both correct, both verified before acting.

P1 -- THE GATE WOULD HAVE FAILED ITS OWN MODULE-SIZE RULE. At 509 lines with
no allowlist row, `node scripts/check-module-size.mjs` exits 1 on this branch.
Confirmed by running it. AGENTS.md says prefer splitting to allowlisting, and
there was already an implicit seam to split on: `evaluateFreshness` in
`lib/base-freshness.mjs` is pure, every git and GitHub call is effects, and the
gate between them only orchestrates. The effects now live in
`lib/base-freshness-io.mjs`; the gate is 339 lines and the gate exits 0.

`isRepoFile` stays with the gate and is passed in, because it is the discovery
predicate rather than an I/O concern. `REPO_ROOT` is computed inside the I/O
module rather than threaded through eight signatures -- every git call there
runs against the repository the file is part of, and a parameter would invite a
caller to point them somewhere else.

The split also exposed a real gap in the tests, which is worth recording rather
than quietly fixing: after the move the suite stayed 16/16 green while the gate
threw `ReferenceError: ghJson is not defined` on its first live run. Every test
drives `--from-json`, which never enters `main()`'s fetch path. That is the
same "the tests do not cover the live path" finding the correctness pass raised
as F3, arriving from a different direction.

P2 -- THE LABEL BOOTSTRAP COULD NOT CREATE ITS LABEL. Applying a label to a PR
is `pull-requests`; CREATING the repository label is `POST /repos/:o/:r/labels`,
which GitHub files under Issues. `base-stale` does not exist in this repo yet
(checked with `gh label list`), so the first sweep has to create it. Both
scopes are now granted.

Measured while checking it, and it changes how load-bearing that step is:
`POST /repos/:o/:r/issues/:n/labels` AUTO-CREATES a missing label. So the sweep
would have limped along without the bootstrap -- but the label would arrive
with no colour and no description. The step stays, and it now PRINTS its error
instead of routing it to /dev/null, since `|| true` on a real permissions
failure is the absence-reads-as-success shape this gate exists to catch.

Verified after the split, on live PRs rather than fixtures: #3578 OK exit 0,
#3610 OK exit 0, 16/16 unit tests, module-size exit 0, test-wiring exit 0.
